refactor + tweaks
