+++ bin/lib/ft/capi.js
