+++ bin/lib/translators/aws.js
@@ -1,7 +1,7 @@
 const AWS = require('aws-sdk');
 AWS.config.update({region: process.env.AWS_REGION || 'eu-west-1'});
 
-const Utils = require('./utils/utils');
+const Utils = require('../utils/utils');
 const BYTE_LIMIT = 5000;
 const THROTTLE_LIMIT = 10*1000 // 10 seconds
